refactor(leantui): extract usage/tool trackers and split update.go
The model struct and update.go had accumulated several distinct concerns.
This splits them into cohesive units without changing behavior:

- Extract per-session usage aggregation into a usageTracker (usage.go),
  replacing four model fields (usageBySession, rootSessionID,
  latestUsageSessionID, sessionStack) with one.
- Extract in-flight tool-call state into a toolTracker (toolstate.go),
  replacing the tools map and toolOrder slice; finish() returns an
  immutable snapshot for the model to commit.
- Move runtime event handling out of update.go into events.go, leaving
  update.go focused on input and command dispatch (761 -> 418 lines).
- Deduplicate run-lifecycle boilerplate via beginRun().
- Deduplicate thinking-level guard/error handling via
  thinkingLevelChangeable() and reportThinkingLevelError().

No functional change; existing tests updated for the new field names.
